Change display of compounds to include addition time and duration
