+++ assays/utils.py
@@ -19,7 +19,8 @@ from .models import (
     AssayDataUpload,
     AssayCompoundInstance,
     AssayInstance,
-    AssaySampleLocation
+    AssaySampleLocation,
+    TIME_CONVERSIONS
 )
 from compounds.models import (
     CompoundSupplier,
@@ -54,16 +55,6 @@ CHIP_FORMATS = ('Chip',)
 
 REPLACED_DATA_POINT_CODE = 'R'
 EXCLUDED_DATA_POINT_CODE = 'X'
-
-# This shouldn't be repeated like so
-# Converts: days -> minutes, hours -> minutes, minutes->minutes
-TIME_CONVERSIONS = [
-    ('day', 1440),
-    ('hour', 60),
-    ('minute', 1)
-]
-
-TIME_CONVERSIONS = collections.OrderedDict(TIME_CONVERSIONS)
 
 # TODO PLEASE REVIEW TO MAKE SURE CONSISTENT
 COLUMN_HEADERS = (
@@ -210,27 +201,6 @@ class UnicodeWriter:
         """This function writes out all rows given"""
         for row in rows:
             self.writerow(row)
-
-
-# TODO EMPLOY THIS FUNCTION ELSEWHERE
-def get_split_times(time_in_minutes):
-    """Takes time_in_minutes and returns a dic with the time split into day, hour, minute"""
-    times = {
-        'day': 0,
-        'hour': 0,
-        'minute': 0
-    }
-    time_in_minutes_remaining = time_in_minutes
-    for time_unit, conversion in TIME_CONVERSIONS.items():
-        initial_time_for_current_field = int(time_in_minutes_remaining / conversion)
-        if initial_time_for_current_field:
-            times[time_unit] = initial_time_for_current_field
-            time_in_minutes_remaining -= initial_time_for_current_field * conversion
-    # Add fractions of minutes if necessary
-    if time_in_minutes_remaining:
-        times['minute'] += time_in_minutes_remaining
-
-    return times
 
 
 def label_to_number(label):
